requirements.txt
tensorflow
sklearn
confluent-kafka
numpy
pandas
json
unicodedata
Pillow
logging
gzip